Separate classes into their own files. Clean up code.
